New coordinator architecture for autocompletion popup. Updated to v0.7.0 of library.

+++ Serifian/Models/Sources/Protocols/AutocompletableSource.swift
@@ -9,5 +9,5 @@ import Foundation
 import SwiftyTypst
 
 protocol AutocompletableSource {
-    func autocomplete(at position: UInt64) -> [AutocompleteResult]
+    func autocomplete(at position: Int) -> [AutocompleteResult]
 }
